pyproject.toml
[tool.poetry]
name = "calorie-tracker"
version = "0.1.0"
description = ""
authors = ["C.Mariappan <c.mariappan@northernarc.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
sqlalchemy = "^2.0.45"
pydantic = "^2.12.5"
python-multipart = "^0.0.21"
requests = "^2.32.5"
pillow = "^12.1.0"
python-dotenv = "^1.2.1"
psycopg2-binary = "^2.9.11"
google-generativeai = "^0.8.6"
google-genai = "^1.57.0"


[tool.poetry.group.dev.dependencies]
black = "^25.12.0"
isort = "^7.0.0"
pytest = "^9.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
